Update CoACD script

+++ data/acronym/convert_mesh_watertight.py
@@ -5,17 +5,19 @@ from pathlib import Path
 from multiprocessing import Process, current_process
 
 import trimesh
+from do_coacd import do_coacd
 
 
 def process_meshes(grasp_h5_paths: list[Path], mesh_dir: Path,
-                   output_dir: Path, script_dir: Path):
+                   output_dir: Path, script_dir: Path = None, coacd_params={}):
     """Run manifold & simplify on meshes and copy corresponding *.mtl files"""
     texture_dir = mesh_dir.parents[1] / "models-textures/textures"
     assert texture_dir.is_dir(), f"{texture_dir=}"
 
-    manifold_path = script_dir / "manifold"
-    simplify_path = script_dir / "simplify"
-    coacd_path = Path(__file__).resolve().parent / "coacd"
+    if script_dir is not None:
+        manifold_path = script_dir / "manifold"
+        simplify_path = script_dir / "simplify"
+    # coacd_path = Path(__file__).resolve().parent / "coacd"
 
     for grasp_h5_path in grasp_h5_paths:
         obj_category, shapenetid, scale = grasp_h5_path.name.split("_")
@@ -39,17 +41,20 @@ def process_meshes(grasp_h5_paths: list[Path], mesh_dir: Path,
         for texture_name in texture_file_names:
             shutil.copy(texture_dir / texture_name, save_path)
 
-        # Process mesh
-        tmp_obj_path = save_path / "temp.watertight.obj"
-        out_mesh_path = save_path / "collision.obj"
-        os.system(f'{manifold_path} {mesh_path} {tmp_obj_path} -s')
-        os.system(f'{simplify_path} -i {tmp_obj_path} -o {out_mesh_path} -m -r 0.02')
-        tmp_obj_path.unlink()
+        # Process mesh (with Manifold)
+        if script_dir is not None:
+            tmp_obj_path = save_path / "temp.watertight.obj"
+            out_mesh_path = save_path / "collision.obj"
+            os.system(f'{manifold_path} {mesh_path} {tmp_obj_path} -s')
+            os.system(f'{simplify_path} -i {tmp_obj_path} -o {out_mesh_path} -m -r 0.02')
+            tmp_obj_path.unlink()
 
         # When failed to process using Manifold, switch to use CoACD
+        out_mesh_path = save_path / "collision.coacd.ply"
         if not out_mesh_path.is_file():
-            os.system(f'{coacd_path} {mesh_path} {out_mesh_path}')
-            os.system(f'touch {save_path / "processed_by_coacd"}')
+            _ = do_coacd(str(mesh_path), outfile=str(out_mesh_path), **coacd_params)
+            # os.system(f'{coacd_path} {mesh_path} {out_mesh_path}')
+            # os.system(f'touch {save_path / "processed_by_coacd"}')
         if not out_mesh_path.is_file():
             print(f"ERROR: Failed for {shapenetid=}")
             continue
@@ -77,7 +82,7 @@ if __name__ == "__main__":
     default_grasp_dir = root_dir / "acronym_grasps"
     default_model_obj_dir = root_dir / "ShapeNetSem-archive/ShapeNetSem-backup/models-OBJ/models"
     default_output_dir = root_dir / "models"
-    default_script_dir = root_dir / "Manifold/build"
+    # default_script_dir = root_dir / "Manifold/build"
 
     parser.add_argument("--grasp-dir", type=str, default=default_grasp_dir,
                         help="Directory of ACRONYM grasps")
@@ -85,8 +90,8 @@ if __name__ == "__main__":
                         help="Directory models-OBJ path")
     parser.add_argument("--output-dir", type=str, default=default_output_dir,
                         help="Output directory path")
-    parser.add_argument("--script-dir", type=str, default=default_script_dir,
-                        help="Directory of scripts to create watertight meshes")
+    # parser.add_argument("--script-dir", type=str, default=default_script_dir,
+    #                     help="Directory of scripts to create watertight meshes")
     parser.add_argument("--num-proc", type=int, default=10,
                         help="Number of parallel processes to start")
 
@@ -96,9 +101,9 @@ if __name__ == "__main__":
     assert grasp_dir.is_dir(), f"{grasp_dir=}"
     mesh_dir = Path(args.mesh_dir)
     assert mesh_dir.is_dir(), f"{mesh_dir=}"
-    script_dir = Path(args.script_dir)
-    assert (script_dir / "manifold").is_file(), f"No manifold in {script_dir=}"
-    assert (script_dir / "simplify").is_file(), f"No simplify in {script_dir=}"
+    # script_dir = Path(args.script_dir)
+    # assert (script_dir / "manifold").is_file(), f"No manifold in {script_dir=}"
+    # assert (script_dir / "simplify").is_file(), f"No simplify in {script_dir=}"
     output_dir = Path(args.output_dir)
     output_dir.mkdir(parents=True, exist_ok=False)
 
@@ -115,7 +120,18 @@ if __name__ == "__main__":
             obj_paths_proc = grasp_obj_paths[num_objs_each*i:num_objs_each*(i+1)]
         processes.append(
             Process(target=process_meshes, name=f"Proc {i}",
-                    args=(obj_paths_proc, mesh_dir, output_dir, script_dir))
+                    args=(
+                        obj_paths_proc,
+                        mesh_dir,
+                        output_dir,
+                        # script_dir
+                    ),
+                    kwargs=dict(coacd_params={
+                        "threshold": 0.03,
+                        "max_convex_hull": 0,
+                        "preprocess_resolution": 100,
+                        "verbose": True,
+                    }))
         )
 
     print(f"Starting {len(processes)} processes")
